Create effective config by defensive copying of original config map
This change allows the user to pass in an immutable config map.  Before,
the user was required to pass in a mutable map that was modified in place
within `configure()`.  First, this behavior was counter-intuitive;  and
second, this behavior was inconsistent to the generic Avro serializers
and deserializers, which do not (need to) modify the user-supplied input
map.